mysite/pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["Belykh Igor"]


[tool.poetry.dependencies]
python = "^3.9"
django-rest-framework = "^0.1.0"
django-debug-toolbar = "^4.2.0"
django-filter = "^23.3"
django-request-logging = "^0.7.5"
drf-spectacular = "^0.26.5"
gunicorn = "^21.2.0"
navigator = "^0.2.2"
pillow = "^10.0.1"
faker = "^19.8.0"
docker = "^6.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
